+++ bot_diffk.py
@@ -6,17 +6,55 @@ from datetime import datetime, timedelta
 SEED = random.randint(0, 1000)
 
 
-def flix_dou_yah(dataset, pe_dim, scenario):
+def flixster(pe_dim, scenario):
     now = (datetime.now() + timedelta(hours=7)).strftime("%b%d_%H:%M")
 
     os.system(
-        f"python Main.py --data-name {dataset} --epochs 40\
-            --testing --ensemble\
+        f"python Main.py\
+            --data-name flixster\
+            --epochs 40\
+            --testing\
+            --ensemble\
             --save-appendix _gatedGCN_{pe_dim}_{scenario}\
             --data-appendix _{pe_dim}\
             --seed {SEED}\
             --pe-dim {pe_dim}\
-            --scenario {scenario} > logs/{dataset}_{scenario}_{pe_dim}_{now}.log"
+            --scenario {scenario} > logs/flixster_{scenario}_{pe_dim}_{now}.log"
+        # --wandb
+    )
+
+
+def douban(pe_dim, scenario):
+    now = (datetime.now() + timedelta(hours=7)).strftime("%b%d_%H:%M")
+
+    os.system(
+        f"python Main.py\
+            --data-name douban\
+            --epochs 40\
+            --testing\
+            --ensemble\
+            --save-appendix _gatedGCN_{pe_dim}_{scenario}\
+            --data-appendix _{pe_dim}\
+            --seed {SEED}\
+            --pe-dim {pe_dim}\
+            --scenario {scenario} > logs/douban_{scenario}_{pe_dim}_{now}.log"
+        # --wandb
+    )
+
+
+def yahoo(pe_dim, scenario):
+    now = (datetime.now() + timedelta(hours=7)).strftime("%b%d_%H:%M")
+
+    os.system(
+        f"python Main.py\
+            --data-name yahoo_music\
+            --epochs 100\
+            --ensemble\
+            --save-appendix _gatedGCN_{pe_dim}_{scenario}\
+            --data-appendix _{pe_dim}\
+            --seed {SEED}\
+            --pe-dim {pe_dim}\
+            --scenario {scenario} > logs/yahoo_{scenario}_{pe_dim}_{now}.log"
         # --wandb
     )
 
@@ -33,7 +71,8 @@ if __name__ == "__main__":
         "flixster": 86,
     }
 
-    for dataset in ["yahoo_music", "douban", "flixster"]:
-        flix_dou_yah(dataset, pe_dims[dataset], args.scenario)
+    flixster(pe_dims["flixster"], args.scenario)
+    douban(pe_dims["douban"], args.scenario)
+    yahoo(pe_dims["yahoo_music"], args.scenario)
 
     # python Main.py --data-name yahoo_music --epochs 40 --testing --ensemble --save-appendix _80 --data-appendix _80 --seed 42 --pe-dim 80