+++ run.sh
@@ -1,9 +1,12 @@
 #/bin/bash
 #douban, yahoo_music
-python Main.py --data-name yahoo_music\
-    --epochs 40\
-    --testing --ensemble\
-    --save-appendix _gatedGCN_20_6\
-    --data-appendix _20\
-    --pe-dim 20\
-    --scenario 6\
+python Main.py\
+        --data-name yahoo_music\
+        --epochs 100\
+        --save-appendix _gatedGCN_20_7\
+        --data-appendix _20\
+        --pe-dim 20\
+        --ensemble\
+        --testing\
+        --scenario 7
+        # --no-train\