+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Mar 18 | 23:05
+
+- Tailor different training scripts for each dataset
+- Add implementation for **RGatedGCNLayer**, and **FastRGatedGCNLayer** (still error)
+
 ## Mar 13 | 12:04
 
 - Add `GatedGCNLSPELayer` and code for scenario [5, 6, 7, 8]